chore: Add kolmafia-js 1.0.3 for dev
I don't know if I'll actually need this as a dependency (instead of a
devDependency). If I ever make types or interfaces from kolmafia into
part of my library's type signature, I will need this as a dependency.
Otherwise, I only need this as a devDependency. But I don't know yet!

So, to be on the safe side, let's add it as a dependency.