build: Transpile to ES5+ with Bublé
Bublé is a fast transpiler that supports a decent level of
customization. Using a custom configuration called `buble-config-rhino`,
we can generate code that takes advantage of ES2015+ features supported
by Rhino, while transpiling features that are not.

Note: Adding Bublé causes Rollup to emit source maps with
`sourcesContent` values of `null`. Not sure if this is OK or not...